Add support for building Swift SDKs for armv7 with target Swift toolchain or a container (#170)
This would implement #65. Although there is no official support for armv7 in Swift yet, this opens the door for being able to generate Swift SDKs for cross compilation to armv7 given you have build artifacts to give it. In my case, I can grab some that I built from my [swift-armv7 repo](https://github.com/xtremekforever/swift-armv7/releases/tag/6.0.3), download and extract, then run the generator like this:

```bash
$ swift run swift-sdk-generator make-linux-sdk --swift-version 6.0.3-RELEASE --target armv7-unknown-linux-gnueabihf --target-swift-package-path ~/Downloads/swift-6.0.3-RELEASE-ubuntu-jammy-armv7-install
...
All done! Install the newly generated SDK with this command:
swift experimental-sdk install ~/swift-sdk-generator/Bundles/6.0.3-RELEASE_ubuntu_jammy_armv7.artifactbundle

After that, use the newly installed SDK when building with this command:
swift build --experimental-swift-sdk 6.0.3-RELEASE_ubuntu_jammy_armv7
```

And, this generates a working Swift SDK that I can use to cross compile anything, like a Hummingbird app:

```bash
$ swift build --swift-sdk 6.0.3-RELEASE_ubuntu_jammy_armv7
warning: 'swift-algorithms': found 1 file(s) which are unhandled; explicitly declare them as resources or exclude from the target
    ~/hummingbird-examples/hello/.build/checkouts/swift-algorithms/Sources/Algorithms/Documentation.docc
Building for debugging...
[948/948] Linking App
Build complete! (35.16s)
```

So, I think that it could be reasonable to add this support to the generator now. It's something I will use myself, but also when official support for armv7 lands for Swift, it'll be ready ;)

+++ Sources/SwiftSDKGenerator/SwiftSDKRecipes/LinuxRecipe.swift
@@ -158,7 +158,7 @@ public struct LinuxRecipe: SwiftSDKRecipe {
       self.linuxDistribution
         .release
     )_\(
-      self.mainTargetTriple.arch!.linuxConventionName
+      self.mainTargetTriple.archName
     )
     """
   }
@@ -217,6 +217,12 @@ public struct LinuxRecipe: SwiftSDKRecipe {
     engine: QueryEngine,
     httpClient client: some HTTPClientProtocol
   ) async throws -> SwiftSDKProduct {
+    if self.linuxDistribution.name == .rhel && self.mainTargetTriple.archName == "armv7" {
+      throw GeneratorError.distributionDoesNotSupportArchitecture(
+        self.linuxDistribution, targetArchName: self.mainTargetTriple.archName
+      )
+    }
+  
     let sdkDirPath = self.sdkDirPath(paths: generator.pathsConfiguration)
     if !generator.isIncremental {
       try await generator.removeRecursively(at: sdkDirPath)
